Add float commands, float motor control that works, basic speed sensing

+++ hexabot.X/app_registers.c
@@ -41,8 +41,8 @@ uint8_t motor_r_dir = 1;
 uint8_t motor_l_pwm = 127;
 uint8_t motor_l_dir = 1;
 
-int8_t r_target_speed = 0;
-int8_t l_target_speed = 0;
+float r_target_speed = 0;
+float l_target_speed = 0;
 
 const register_t registers[] = {
     // 0
